+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/AbstractParallelEoSStreamProcessor.java
@@ -5,14 +5,16 @@ package io.confluent.parallelconsumer.internal;
  */
 
 import io.confluent.csid.utils.TimeUtils;
-import io.confluent.parallelconsumer.*;
+import io.confluent.parallelconsumer.ParallelConsumer;
+import io.confluent.parallelconsumer.ParallelConsumerException;
+import io.confluent.parallelconsumer.ParallelConsumerOptions;
+import io.confluent.parallelconsumer.PollContextInternal;
 import io.confluent.parallelconsumer.state.WorkManager;
 import io.micrometer.core.instrument.Timer;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.Setter;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
@@ -36,12 +38,9 @@ import java.util.function.Function;
 import java.util.regex.Pattern;
 
 import static io.confluent.csid.utils.BackportUtils.isEmpty;
-import static io.confluent.csid.utils.BackportUtils.toSeconds;
-import static io.confluent.csid.utils.StringUtils.msg;
 import static io.confluent.parallelconsumer.internal.State.*;
 import static java.lang.Boolean.TRUE;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static lombok.AccessLevel.PROTECTED;
 import static lombok.AccessLevel.PUBLIC;
 
@@ -49,7 +48,7 @@ import static lombok.AccessLevel.PUBLIC;
  * @see ParallelConsumer
  */
 @Slf4j
-public abstract class AbstractParallelEoSStreamProcessor<K, V> implements ParallelConsumer<K, V>, ConsumerRebalanceListener, Closeable {
+public abstract class AbstractParallelEoSStreamProcessor<K, V> implements ParallelConsumer<K, V>, Closeable {
 
     public static final String MDC_INSTANCE_ID = "pcId";
 
@@ -319,61 +318,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
     }
 
     /**
-     * Commit our offsets
-     * <p>
-     * Make sure the calling thread is the thread which performs commit - i.e. is the {@link OffsetCommitter}.
-     */
-    @Override
-    public void onPartitionsRevoked(Collection<TopicPartition> partitions) {
-        log.debug("Partitions revoked {}, state: {}", partitions, state);
-        numberOfAssignedPartitions = numberOfAssignedPartitions - partitions.size();
-
-        try {
-            // commit any offsets from revoked partitions BEFORE truncation
-            commitOffsetsThatAreReady();
-
-            // truncate the revoked partitions
-            wm.onPartitionsRevoked(partitions);
-        } catch (Exception e) {
-            throw new InternalRuntimeException("onPartitionsRevoked event error", e);
-        }
-
-        //
-        try {
-            usersConsumerRebalanceListener.ifPresent(listener -> listener.onPartitionsRevoked(partitions));
-        } catch (Exception e) {
-            throw new ExceptionInUserFunctionException("Error from rebalance listener function after #onPartitionsRevoked", e);
-        }
-    }
-
-    /**
-     * Delegate to {@link WorkManager}
-     *
-     * @see WorkManager#onPartitionsAssigned
-     */
-    @Override
-    public void onPartitionsAssigned(Collection<TopicPartition> partitions) {
-        numberOfAssignedPartitions = numberOfAssignedPartitions + partitions.size();
-        log.info("Assigned {} total ({} new) partition(s) {}", numberOfAssignedPartitions, partitions.size(), partitions);
-        wm.onPartitionsAssigned(partitions);
-        usersConsumerRebalanceListener.ifPresent(x -> x.onPartitionsAssigned(partitions));
-        notifySomethingToDo();
-    }
-
-    /**
-     * Cannot commit any offsets for partitions that have been `lost` (as opposed to revoked). Just delegate to
-     * {@link WorkManager} for truncation.
-     *
-     * @see WorkManager#onPartitionsAssigned
-     */
-    @Override
-    public void onPartitionsLost(Collection<TopicPartition> partitions) {
-        numberOfAssignedPartitions = numberOfAssignedPartitions - partitions.size();
-        wm.onPartitionsLost(partitions);
-        usersConsumerRebalanceListener.ifPresent(x -> x.onPartitionsLost(partitions));
-    }
-
-    /**
      * Nasty reflection to check if auto commit is disabled.
      * <p>
      * Other way would be to politely request the user also include their consumer properties when construction, but
@@ -408,132 +352,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         }
     }
 
-    /**
-     * Close the system, without draining.
-     *
-     * @see State#draining
-     */
-    @Override
-    public void close() {
-        // use a longer timeout, to cover for evey other step using the default
-        Duration timeout = DrainingCloseable.DEFAULT_TIMEOUT.multipliedBy(2);
-        closeDontDrainFirst(timeout);
-    }
-
-    @Override
-    @SneakyThrows
-    public void close(Duration timeout, DrainingMode drainMode) {
-        if (state == closed) {
-            log.info("Already closed, checking end state..");
-        } else {
-            log.info("Signaling to close...");
-
-            switch (drainMode) {
-                case DRAIN -> {
-                    log.info("Will wait for all in flight to complete before");
-                    transitionToDraining();
-                }
-                case DONT_DRAIN -> {
-                    log.info("Not waiting for remaining queued to complete, will finish in flight, then close...");
-                    transitionToClosing();
-                }
-            }
-
-            waitForClose(timeout);
-        }
-
-        if (controlThreadFuture.isPresent()) {
-            log.debug("Checking for control thread exception...");
-            Future<?> future = controlThreadFuture.get();
-            future.get(timeout.toMillis(), MILLISECONDS); // throws exception if supervisor saw one
-        }
-
-        log.info("Close complete.");
-    }
-
-    private void waitForClose(Duration timeout) throws TimeoutException, ExecutionException {
-        log.info("Waiting on closed state...");
-        while (!state.equals(closed)) {
-            try {
-                Future<Boolean> booleanFuture = this.controlThreadFuture.get();
-                log.debug("Blocking on control future");
-                boolean signaled = booleanFuture.get(toSeconds(timeout), SECONDS);
-                if (!signaled)
-                    throw new TimeoutException("Timeout waiting for system to close (" + timeout + ")");
-            } catch (InterruptedException e) {
-                // ignore
-                log.trace("Interrupted", e);
-            } catch (ExecutionException | TimeoutException e) {
-                log.error("Execution or timeout exception while waiting for the control thread to close cleanly " +
-                        "(state was {}). Try increasing your time-out to allow the system to drain, or close without " +
-                        "draining.", state, e);
-                throw e;
-            }
-            log.trace("Still waiting for system to close...");
-        }
-    }
-
-    private void doClose(Duration timeout) throws TimeoutException, ExecutionException, InterruptedException {
-        log.debug("Starting close process (state: {})...", state);
-
-        log.debug("Shutting down execution pool...");
-        List<Runnable> unfinished = workerThreadPool.shutdownNow();
-        if (!unfinished.isEmpty()) {
-            log.warn("Threads not done count: {}", unfinished.size());
-        }
-
-        log.debug("Awaiting worker pool termination...");
-        boolean interrupted = true;
-        while (interrupted) {
-            log.debug("Still interrupted");
-            try {
-                boolean terminationFinishedWithoutTimeout = workerThreadPool.awaitTermination(toSeconds(timeout), SECONDS);
-                interrupted = false;
-                if (!terminationFinishedWithoutTimeout) {
-                    log.warn("Thread execution pool termination await timeout ({})! Were any processing jobs dead locked (test latch locks?) or otherwise stuck?", timeout);
-                }
-            } catch (InterruptedException e) {
-                log.error("InterruptedException", e);
-                interrupted = true;
-            }
-        }
-        log.debug("Worker pool terminated.");
-
-        // last check to see if after worker pool closed, has any new work arrived?
-        workMailbox.processWorkCompleteMailBox(Duration.ZERO);
-
-        //
-        commitOffsetsThatAreReady();
-
-        // only close consumer once producer has committed it's offsets (tx'l)
-        log.debug("Closing and waiting for broker poll system...");
-        brokerPollSubsystem.closeAndWait();
-
-        maybeCloseConsumer();
-
-        producerManager.ifPresent(x -> x.close(timeout));
-
-        log.debug("Close complete.");
-        this.state = closed;
-
-        if (this.getFailureCause() != null) {
-            log.error("PC closed due to error: {}", getFailureCause(), null);
-        }
-    }
-
-    /**
-     * To keep things simple, make sure the correct thread which can make a commit, is the one to close the consumer.
-     * This way, if partitions are revoked, the commit can be made inline.
-     */
-    private void maybeCloseConsumer() {
-        if (isResponsibleForCommits()) {
-            consumer.close();
-        }
-    }
-
-    private boolean isResponsibleForCommits() {
-        return (committer instanceof ProducerManager);
-    }
 
     private boolean isRecordsAwaitingProcessing() {
         boolean isRecordsAwaitingProcessing = wm.isRecordsAwaitingProcessing();
@@ -542,11 +360,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         return isRecordsAwaitingProcessing || threadsDone;
     }
 
-    private void transitionToDraining() {
-        log.debug("Transitioning to draining...");
-        this.state = State.draining;
-        notifySomethingToDo();
-    }
 
     /**
      * Control thread can be blocked waiting for work, but is interruptible. Interrupting it can be useful to inform
@@ -569,65 +382,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         }
     }
 
-    /**
-     * Kicks off the control loop in the executor, with supervision and returns.
-     *
-     * @see #supervisorLoop(Function, Consumer)
-     */
-    protected <R> void supervisorLoop(Function<PollContextInternal<K, V>, List<R>> userFunctionWrapped,
-                                      Consumer<R> callback) {
-        if (state == State.unused) {
-            state = running;
-        } else {
-            throw new IllegalStateException(msg("Invalid state - you cannot call the poll* or pollAndProduce* methods " +
-                    "more than once (they are asynchronous) (current state is {})", state));
-        }
-
-        // broker poll subsystem
-        brokerPollSubsystem.start(options.getManagedExecutorService());
-
-        ExecutorService executorService;
-        try {
-            executorService = InitialContext.doLookup(options.getManagedExecutorService());
-        } catch (NamingException e) {
-            log.debug("Using Java SE Thread", e);
-            executorService = Executors.newSingleThreadExecutor();
-        }
-
-
-        // run main pool loop in thread
-        Callable<Boolean> controlTask = () -> superviseControlLoop((Function) userFunctionWrapped, (Consumer<Object>) callback);
-        Future<Boolean> controlTaskFutureResult = executorService.submit(controlTask);
-        this.controlThreadFuture = Optional.of(controlTaskFutureResult);
-    }
-
-    // todo name
-    private <R> boolean superviseControlLoop(Function userFunctionWrapped, Consumer<Object> callback) throws Exception {
-        addInstanceMDC(options);
-        log.info("Control loop starting up...");
-
-        initWorkerPool(userFunctionWrapped, callback);
-
-        Thread controlThread = Thread.currentThread();
-        controlThread.setName("pc-control");
-        this.blockableControlThread = controlThread;
-        while (state != closed) {
-            log.debug("Control loop start");
-            try {
-                controlLoop();
-            } catch (InterruptedException e) {
-                log.debug("Control loop interrupted, closing");
-                doClose(DrainingCloseable.DEFAULT_TIMEOUT);
-            } catch (Exception e) {
-                log.error("Error from poll control thread, will attempt controlled shutdown, then rethrow. Error: " + e.getMessage(), e);
-                failureReason = new RuntimeException("Error from poll control thread: " + e.getMessage(), e);
-                doClose(DrainingCloseable.DEFAULT_TIMEOUT); // attempt to close
-                throw failureReason;
-            }
-        }
-        log.info("Control loop ending clean (state:{})...", state);
-        return true;
-    }
 
     private static void initWorkerPool(Function userFunctionWrapped, Consumer<Object> callback) {
         // todo casts
@@ -650,57 +404,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         options.getMyId().ifPresent(id -> MDC.put(MDC_INSTANCE_ID, id));
     }
 
-    /**
-     * Main control loop
-     */
-    protected <R> void controlLoop() throws TimeoutException, ExecutionException, InterruptedException {
-        maybeWakeupPoller();
-
-        //
-        final boolean shouldTryCommitNow = maybeAcquireCommitLock();
-
-        // make sure all work that's been completed are arranged ready for commit
-        Duration timeToBlockFor = shouldTryCommitNow ? Duration.ZERO : getTimeToBlockFor();
-        workMailbox.processWorkCompleteMailBox(timeToBlockFor);
-
-        //
-        if (shouldTryCommitNow) {
-            // offsets will be committed when the consumer has its partitions revoked
-            commitOffsetsThatAreReady();
-        }
-
-        // distribute more work
-        retrieveAndDistributeNewWorkNew();
-
-        // run call back
-        log.trace("Loop: Running {} loop end plugin(s)", controlLoopHooks.size());
-        this.controlLoopHooks.forEach(Runnable::run);
-
-        log.trace("Current state: {}", state);
-        switch (state) {
-            case draining -> {
-                drain();
-            }
-            case closing -> {
-                doClose(DrainingCloseable.DEFAULT_TIMEOUT);
-            }
-        }
-
-        // sanity - supervise the poller
-        brokerPollSubsystem.supervise();
-
-        // thread yield for spin lock avoidance
-        Duration duration = Duration.ofMillis(1);
-        try {
-            Thread.sleep(duration.toMillis());
-        } catch (InterruptedException e) {
-            log.trace("Woke up", e);
-        }
-
-        // end of loop
-        log.trace("End of control loop, waiting processing {}, remaining in partition queues: {}, out for processing: {}. In state: {}",
-                wm.getNumberOfWorkQueuedInShardsAwaitingSelection(), wm.getNumberOfIncompleteOffsets(), wm.getNumberRecordsOutForProcessing(), state);
-    }
 
     /**
      * If we don't have enough work queued, and the poller is paused for throttling,
@@ -745,91 +448,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         return work.size();
     }
 
-//    private <R> int retrieveAndDistributeNewWorkOld(final Function<PollContextInternal<K, V>, List<R>> userFunction, final Consumer<R> callback) {
-//        // check queue pressure first before addressing it
-//        checkPipelinePressure();
-//
-//        int gotWorkCount = 0;
-//
-//        //
-//        if (state == running || state == draining) {
-//            int delta = calculateQuantityToRequest();
-//            var records = wm.getWorkIfAvailable(delta);
-//
-//            gotWorkCount = records.size();
-//            lastWorkRequestWasFulfilled = gotWorkCount >= delta;
-//
-//            log.trace("Loop: Submit to pool");
-//            submitWorkToPool(userFunction, callback, records);
-//        }
-//
-//        //
-//        queueStatsLimiter.performIfNotLimited(() -> {
-//            int queueSize = getNumberOfUserFunctionsQueued();
-//            log.debug("Stats: \n- pool active: {} queued:{} \n- queue size: {} target: {} loading factor: {}",
-//                    workerThreadPool.getActiveCount(), queueSize, queueSize, getPoolLoadTarget(), dynamicExtraLoadFactor.getCurrentFactor());
-//        });
-//
-//        return gotWorkCount;
-//    }
-
-//    /**
-//     * @return number of {@link WorkContainer} to try to get
-//     */
-//    protected int calculateQuantityToRequest() {
-//        int target = getTargetOutForProcessing();
-//        int current = wm.getNumberRecordsOutForProcessing();
-//        int delta = target - current;
-//
-//        // always round up to fill batches - get however extra are needed to fill a batch
-//        if (options.isUsingBatching()) {
-//            //noinspection OptionalGetWithoutIsPresent
-//            int batchSize = options.getBatchSize();
-//            int modulo = delta % batchSize;
-//            if (modulo > 0) {
-//                int extraToFillBatch = target - modulo;
-//                delta = delta + extraToFillBatch;
-//            }
-//        }
-//
-//        log.debug("Will try to get work - target: {}, current queue size: {}, requesting: {}, loading factor: {}",
-//                target, current, delta, dynamicExtraLoadFactor.getCurrentFactor());
-//        return delta;
-//    }
-
-//    protected int getTargetOutForProcessing() {
-//        return getQueueTargetLoaded();
-//    }
-
-//    protected int getQueueTargetLoaded() {
-//        //noinspection unchecked
-//        int batch = options.getBatchSize();
-//        return getPoolLoadTarget() * dynamicExtraLoadFactor.getCurrentFactor() * batch;
-//    }
-
-//    /**
-//     * Checks the system has enough pressure in the pipeline of work, if not attempts to step up the load factor.
-//     */
-//    protected void checkPipelinePressure() {
-//        if (log.isTraceEnabled())
-//            log.trace("Queue pressure check: (current size: {}, loaded target: {}, factor: {}) " +
-//                            "if (isPoolQueueLow() {} && lastWorkRequestWasFulfilled {}))",
-//                    getNumberOfUserFunctionsQueued(),
-//                    getQueueTargetLoaded(),
-//                    dynamicExtraLoadFactor.getCurrentFactor(),
-//                    isPoolQueueLow(),
-//                    lastWorkRequestWasFulfilled);
-//
-//        if (isPoolQueueLow() && lastWorkRequestWasFulfilled) {
-//            boolean steppedUp = dynamicExtraLoadFactor.maybeStepUp();
-//            if (steppedUp) {
-//                log.debug("isPoolQueueLow(): Executor pool queue is not loaded with enough work (queue: {} vs target: {}), stepped up loading factor to {}",
-//                        getNumberOfUserFunctionsQueued(), getPoolLoadTarget(), dynamicExtraLoadFactor.getCurrentFactor());
-//            } else if (dynamicExtraLoadFactor.isMaxReached()) {
-//                log.warn("isPoolQueueLow(): Max loading factor steps reached: {}/{}", dynamicExtraLoadFactor.getCurrentFactor(), dynamicExtraLoadFactor.getMaxFactor());
-//            }
-//        }
-//    }
 
     /**
      * @return aim to never have the pool queue drop below this
@@ -838,34 +456,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         return options.getTargetAmountOfRecordsInFlight();
     }
 
-//    private boolean isPoolQueueLow() {
-//        int queueSize = getNumberOfUserFunctionsQueued();
-//        int queueTarget = getPoolLoadTarget();
-//        boolean workAmountBelowTarget = queueSize <= queueTarget;
-//        log.debug("isPoolQueueLow()? workAmountBelowTarget {} {} vs {};",
-//                workAmountBelowTarget, queueSize, queueTarget);
-//        return workAmountBelowTarget;
-//    }
-
-    private void drain() {
-        log.debug("Signaling to drain...");
-        brokerPollSubsystem.drain();
-        if (!isRecordsAwaitingProcessing()) {
-            transitionToClosing();
-        } else {
-            log.debug("Records still waiting processing, won't transition to closing.");
-        }
-    }
-
-    private void transitionToClosing() {
-        log.debug("Transitioning to closing...");
-        if (state == State.unused) {
-            state = closed;
-        } else {
-            state = State.closing;
-        }
-        notifySomethingToDo();
-    }
 
     /**
      * The amount of time to block poll in this cycle
