+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/Controller.java
@@ -1,0 +1,294 @@
+package io.confluent.parallelconsumer.internal;
+
+import io.confluent.parallelconsumer.PollContextInternal;
+import lombok.SneakyThrows;
+
+import javax.naming.InitialContext;
+import javax.naming.NamingException;
+import java.time.Duration;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.*;
+import java.util.function.Consumer;
+import java.util.function.Function;
+
+import static io.confluent.csid.utils.BackportUtils.toSeconds;
+import static io.confluent.parallelconsumer.internal.State.closed;
+import static io.confluent.parallelconsumer.internal.State.running;
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
+import static java.util.concurrent.TimeUnit.SECONDS;
+
+/**
+ * @author Antony Stubbs
+ */
+public class Controller {
+
+
+    /**
+     * Kicks off the control loop in the executor, with supervision and returns.
+     *
+     * @see #supervisorLoop(Function, Consumer)
+     */
+    protected <R> void supervisorLoop(Function<PollContextInternal<K, V>, List<R>> userFunctionWrapped,
+                                      Consumer<R> callback) {
+        if (state == State.unused) {
+            state = running;
+        } else {
+            throw new IllegalStateException(msg("Invalid state - you cannot call the poll* or pollAndProduce* methods " +
+                    "more than once (they are asynchronous) (current state is {})", state));
+        }
+
+        // broker poll subsystem
+        brokerPollSubsystem.start(options.getManagedExecutorService());
+
+        ExecutorService executorService;
+        try {
+            executorService = InitialContext.doLookup(options.getManagedExecutorService());
+        } catch (NamingException e) {
+            log.debug("Using Java SE Thread", e);
+            executorService = Executors.newSingleThreadExecutor();
+        }
+
+
+        // run main pool loop in thread
+        Callable<Boolean> controlTask = () -> superviseControlLoop((Function) userFunctionWrapped, (Consumer<Object>) callback);
+        Future<Boolean> controlTaskFutureResult = executorService.submit(controlTask);
+        this.controlThreadFuture = Optional.of(controlTaskFutureResult);
+    }
+
+    // todo name
+    private <R> boolean superviseControlLoop(Function userFunctionWrapped, Consumer<Object> callback) throws Exception {
+        addInstanceMDC(options);
+        log.info("Control loop starting up...");
+
+        initWorkerPool(userFunctionWrapped, callback);
+
+        Thread controlThread = Thread.currentThread();
+        controlThread.setName("pc-control");
+        this.blockableControlThread = controlThread;
+        while (state != closed) {
+            log.debug("Control loop start");
+            try {
+                controlLoop();
+            } catch (InterruptedException e) {
+                log.debug("Control loop interrupted, closing");
+                doClose(DrainingCloseable.DEFAULT_TIMEOUT);
+            } catch (Exception e) {
+                log.error("Error from poll control thread, will attempt controlled shutdown, then rethrow. Error: " + e.getMessage(), e);
+                failureReason = new RuntimeException("Error from poll control thread: " + e.getMessage(), e);
+                doClose(DrainingCloseable.DEFAULT_TIMEOUT); // attempt to close
+                throw failureReason;
+            }
+        }
+        log.info("Control loop ending clean (state:{})...", state);
+        return true;
+    }
+
+
+    /**
+     * Main control loop
+     */
+    protected void controlLoop() throws TimeoutException, ExecutionException, InterruptedException {
+        initWorkerPool();
+        maybeWakeupPoller();
+
+        //
+        final boolean shouldTryCommitNow = maybeAcquireCommitLock();
+
+        // make sure all work that's been completed are arranged ready for commit
+        Duration timeToBlockFor = shouldTryCommitNow ? Duration.ZERO : getTimeToBlockFor();
+        workMailbox.processWorkCompleteMailBox(timeToBlockFor);
+
+        //
+        if (shouldTryCommitNow) {
+            // offsets will be committed when the consumer has its partitions revoked
+            commitOffsetsThatAreReady();
+        }
+
+        // distribute more work
+        retrieveAndDistributeNewWorkNew();
+
+        // run call back
+        log.trace("Loop: Running {} loop end plugin(s)", controlLoopHooks.size());
+        this.controlLoopHooks.forEach(Runnable::run);
+
+        log.trace("Current state: {}", state);
+        switch (state) {
+            case draining -> {
+                drain();
+            }
+            case closing -> {
+                doClose(DrainingCloseable.DEFAULT_TIMEOUT);
+            }
+        }
+
+        // sanity - supervise the poller
+        brokerPollSubsystem.supervise();
+
+        // thread yield for spin lock avoidance
+        Duration duration = Duration.ofMillis(1);
+        try {
+            Thread.sleep(duration.toMillis());
+        } catch (InterruptedException e) {
+            log.trace("Woke up", e);
+        }
+
+        // end of loop
+        log.trace("End of control loop, waiting processing {}, remaining in partition queues: {}, out for processing: {}. In state: {}",
+                wm.getNumberOfWorkQueuedInShardsAwaitingSelection(), wm.getNumberOfIncompleteOffsets(), wm.getNumberRecordsOutForProcessing(), state);
+    }
+
+    /**
+     * Close the system, without draining.
+     *
+     * @see State#draining
+     */
+    @Override
+    public void close() {
+        // use a longer timeout, to cover for evey other step using the default
+        Duration timeout = DrainingCloseable.DEFAULT_TIMEOUT.multipliedBy(2);
+        closeDontDrainFirst(timeout);
+    }
+
+    @Override
+    @SneakyThrows
+    public void close(Duration timeout, DrainingCloseable.DrainingMode drainMode) {
+        if (state == closed) {
+            log.info("Already closed, checking end state..");
+        } else {
+            log.info("Signaling to close...");
+
+            switch (drainMode) {
+                case DRAIN -> {
+                    log.info("Will wait for all in flight to complete before");
+                    transitionToDraining();
+                }
+                case DONT_DRAIN -> {
+                    log.info("Not waiting for remaining queued to complete, will finish in flight, then close...");
+                    transitionToClosing();
+                }
+            }
+
+            waitForClose(timeout);
+        }
+
+        if (controlThreadFuture.isPresent()) {
+            log.debug("Checking for control thread exception...");
+            Future<?> future = controlThreadFuture.get();
+            future.get(timeout.toMillis(), MILLISECONDS); // throws exception if supervisor saw one
+        }
+
+        log.info("Close complete.");
+    }
+
+    private void waitForClose(Duration timeout) throws TimeoutException, ExecutionException {
+        log.info("Waiting on closed state...");
+        while (!state.equals(closed)) {
+            try {
+                Future<Boolean> booleanFuture = this.controlThreadFuture.get();
+                log.debug("Blocking on control future");
+                boolean signaled = booleanFuture.get(toSeconds(timeout), SECONDS);
+                if (!signaled)
+                    throw new TimeoutException("Timeout waiting for system to close (" + timeout + ")");
+            } catch (InterruptedException e) {
+                // ignore
+                log.trace("Interrupted", e);
+            } catch (ExecutionException | TimeoutException e) {
+                log.error("Execution or timeout exception while waiting for the control thread to close cleanly " +
+                        "(state was {}). Try increasing your time-out to allow the system to drain, or close without " +
+                        "draining.", state, e);
+                throw e;
+            }
+            log.trace("Still waiting for system to close...");
+        }
+    }
+
+    private void doClose(Duration timeout) throws TimeoutException, ExecutionException, InterruptedException {
+        log.debug("Starting close process (state: {})...", state);
+
+        log.debug("Shutting down execution pool...");
+        List<Runnable> unfinished = workerThreadPool.shutdownNow();
+        if (!unfinished.isEmpty()) {
+            log.warn("Threads not done count: {}", unfinished.size());
+        }
+
+        log.debug("Awaiting worker pool termination...");
+        boolean interrupted = true;
+        while (interrupted) {
+            log.debug("Still interrupted");
+            try {
+                boolean terminationFinishedWithoutTimeout = workerThreadPool.awaitTermination(toSeconds(timeout), SECONDS);
+                interrupted = false;
+                if (!terminationFinishedWithoutTimeout) {
+                    log.warn("Thread execution pool termination await timeout ({})! Were any processing jobs dead locked (test latch locks?) or otherwise stuck?", timeout);
+                }
+            } catch (InterruptedException e) {
+                log.error("InterruptedException", e);
+                interrupted = true;
+            }
+        }
+        log.debug("Worker pool terminated.");
+
+        // last check to see if after worker pool closed, has any new work arrived?
+        workMailbox.processWorkCompleteMailBox(Duration.ZERO);
+
+        //
+        commitOffsetsThatAreReady();
+
+        // only close consumer once producer has committed it's offsets (tx'l)
+        log.debug("Closing and waiting for broker poll system...");
+        brokerPollSubsystem.closeAndWait();
+
+        maybeCloseConsumer();
+
+        producerManager.ifPresent(x -> x.close(timeout));
+
+        log.debug("Close complete.");
+        this.state = closed;
+
+        if (this.getFailureCause() != null) {
+            log.error("PC closed due to error: {}", getFailureCause(), null);
+        }
+    }
+
+    /**
+     * To keep things simple, make sure the correct thread which can make a commit, is the one to close the consumer.
+     * This way, if partitions are revoked, the commit can be made inline.
+     */
+    private void maybeCloseConsumer() {
+        if (isResponsibleForCommits()) {
+            consumer.close();
+        }
+    }
+
+    private boolean isResponsibleForCommits() {
+        return (committer instanceof ProducerManager);
+    }
+
+    private void transitionToDraining() {
+        log.debug("Transitioning to draining...");
+        this.state = State.draining;
+        notifySomethingToDo();
+    }
+
+
+    private void drain() {
+        log.debug("Signaling to drain...");
+        brokerPollSubsystem.drain();
+        if (!isRecordsAwaitingProcessing()) {
+            transitionToClosing();
+        } else {
+            log.debug("Records still waiting processing, won't transition to closing.");
+        }
+    }
+
+    private void transitionToClosing() {
+        log.debug("Transitioning to closing...");
+        if (state == State.unused) {
+            state = closed;
+        } else {
+            state = State.closing;
+        }
+        notifySomethingToDo();
+    }
+}
