+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/RebalanceHandler.java
@@ -1,0 +1,67 @@
+package io.confluent.parallelconsumer.internal;
+
+import io.confluent.parallelconsumer.ExceptionInUserFunctionException;
+import io.confluent.parallelconsumer.state.WorkManager;
+import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
+import org.apache.kafka.common.TopicPartition;
+
+import java.util.Collection;
+
+public class RebalanceHandler implements ConsumerRebalanceListener {
+
+    /**
+     * Commit our offsets
+     * <p>
+     * Make sure the calling thread is the thread which performs commit - i.e. is the {@link OffsetCommitter}.
+     */
+    @Override
+    public void onPartitionsRevoked(Collection<TopicPartition> partitions) {
+        log.debug("Partitions revoked {}, state: {}", partitions, state);
+        numberOfAssignedPartitions = numberOfAssignedPartitions - partitions.size();
+
+        try {
+            // commit any offsets from revoked partitions BEFORE truncation
+            commitOffsetsThatAreReady();
+
+            // truncate the revoked partitions
+            wm.onPartitionsRevoked(partitions);
+        } catch (Exception e) {
+            throw new InternalRuntimeException("onPartitionsRevoked event error", e);
+        }
+
+        //
+        try {
+            usersConsumerRebalanceListener.ifPresent(listener -> listener.onPartitionsRevoked(partitions));
+        } catch (Exception e) {
+            throw new ExceptionInUserFunctionException("Error from rebalance listener function after #onPartitionsRevoked", e);
+        }
+    }
+
+    /**
+     * Delegate to {@link WorkManager}
+     *
+     * @see WorkManager#onPartitionsAssigned
+     */
+    @Override
+    public void onPartitionsAssigned(Collection<TopicPartition> partitions) {
+        numberOfAssignedPartitions = numberOfAssignedPartitions + partitions.size();
+        log.info("Assigned {} total ({} new) partition(s) {}", numberOfAssignedPartitions, partitions.size(), partitions);
+        wm.onPartitionsAssigned(partitions);
+        usersConsumerRebalanceListener.ifPresent(x -> x.onPartitionsAssigned(partitions));
+        notifySomethingToDo();
+    }
+
+    /**
+     * Cannot commit any offsets for partitions that have been `lost` (as opposed to revoked). Just delegate to
+     * {@link WorkManager} for truncation.
+     *
+     * @see WorkManager#onPartitionsAssigned
+     */
+    @Override
+    public void onPartitionsLost(Collection<TopicPartition> partitions) {
+        numberOfAssignedPartitions = numberOfAssignedPartitions - partitions.size();
+        wm.onPartitionsLost(partitions);
+        usersConsumerRebalanceListener.ifPresent(x -> x.onPartitionsLost(partitions));
+    }
+
+}
